refactor(frontend): extract abortableSleep to shared helpers
