+++ apps/frontend/src/pages/chat/hooks/useStreamChat.ts
@@ -7,6 +7,7 @@ import {
   subscribeEvents,
 } from '@/api/chat/index.js';
 import {HttpError} from '@/api/helpers/http-error.js';
+import {abortableSleep} from '@/helpers/abortable-sleep.js';
 import {EventBus} from '@/helpers/event-bus.js';
 
 import type {ChatEventMap} from '../components/StreamingMessageDisplay/index.js';
@@ -232,26 +233,4 @@ function isRetriableError(e: unknown): boolean {
   if (e instanceof TypeError) return true;
   if (e instanceof HttpError && e.status >= 500) return true;
   return false;
-}
-
-/**
- * Returns a promise that resolves to `true` after {@link ms} milliseconds,
- * or `false` if the {@link signal} is aborted before the timer fires.
- */
-function abortableSleep(ms: number, signal: AbortSignal): Promise<boolean> {
-  return new Promise((resolve) => {
-    if (signal.aborted) {
-      resolve(false);
-      return;
-    }
-    const timer = setTimeout(() => {
-      signal.removeEventListener('abort', onAbort);
-      resolve(true);
-    }, ms);
-    const onAbort = () => {
-      clearTimeout(timer);
-      resolve(false);
-    };
-    signal.addEventListener('abort', onAbort, {once: true});
-  });
 }
